+++ changelog.md
@@ -2,13 +2,15 @@
 
 ## Todo
 
-- correction filtre + recherche
 
 
 ## Versions
 
 **0.2**
 
+- ajout d'un icône filtre dans le header à droite. Cohérence à apporter avec les filtres existants
+- mise en forme image livre
+- correction filtre + recherche
 - sauvegarde d'un livre en PouchDB (base de données NoSQL locale)
 - modification d'un livre pour le mettre en favori ou non
 - page de debug de l'app